Notify he user when choses a configuration and default configuration
